Created AuthenticationController and cleaned up User Controller.
