+++ Sources/App/Controllers/UserController.swift
@@ -46,31 +46,7 @@ struct UserController: RouteCollection {
         return try await user.dto(on: req.db)
     }
     
-    /// Sign user up.
-    func signUserUp(req: Request) async throws -> User {
-        try User.Create.validate(content: req)
-        let create = try req.content.decode(User.Create.self)
-        guard create.password == create.confirmPassword else {
-            throw Abort(.badRequest, reason: "Passwords did not match.")
-        }
-        
-        let user = try User(
-            email: create.email,
-            passwordHash: Bcrypt.hash(create.password),
-            firstName: "",
-            lastName: "",
-            discordUsername: "",
-            githubUsername: "",
-            tags: [],
-            links: [],
-            profileImage: "",
-            biography: "",
-            location: ""
-        )
-        
-        try await user.save(on: req.db)
-        return user
-    }
+ 
     
     /// Find the User for the provided User ID
     func find(req: Request) async throws -> UserDTO {
